+++ admin/class-device-detector-admin.php
@@ -21,6 +21,7 @@ use PODeviceDetector\System\Date;
 use PODeviceDetector\System\Timezone;
 use PODeviceDetector\Plugin\Feature\CSSModifier;
 use PerfOpsOne\Menus;
+use PerfOpsOne\AdminBar;
 
 /**
  * The admin-specific functionality of the plugin.
@@ -161,6 +162,7 @@ class Device_Detector_Admin {
 	public function init_admin_menus() {
 		add_filter( 'init_perfopsone_admin_menus', [ $this, 'init_perfopsone_admin_menus' ] );
 		Menus::initialize();
+		AdminBar::initialize();
 	}
 
 	/**
@@ -284,6 +286,16 @@ class Device_Detector_Admin {
 									$this->save_options();
 								} elseif ( ! empty( $_POST ) && array_key_exists( 'reset-to-defaults', $_POST ) ) {
 									$this->reset_options();
+								}
+							}
+							break;
+						case 'install-decalog':
+							if ( class_exists( 'PerfOpsOne\Installer' ) ) {
+								$result = \PerfOpsOne\Installer::do( 'decalog', true );
+								if ( '' === $result ) {
+									add_settings_error( 'podd_no_error', '', esc_html__( 'Plugin successfully installed and activated with default settings.', 'device-detector' ), 'info' );
+								} else {
+									add_settings_error( 'podd_install_error', '', sprintf( esc_html__( 'Unable to install or activate the plugin. Error message: %s.', 'device-detector' ), $result ), 'error' );
 								}
 							}
 							break;
@@ -436,6 +448,9 @@ class Device_Detector_Admin {
 		} else {
 			$help  = '<img style="width:16px;vertical-align:text-bottom;" src="' . \Feather\Icons::get_base64( 'alert-triangle', 'none', '#FF8C00' ) . '" />&nbsp;';
 			$help .= sprintf( esc_html__('Your site does not use any logging plugin. To log all events triggered in Device Detector, I recommend you to install the excellent (and free) %s. But it is not mandatory.', 'device-detector' ), '<a href="https://wordpress.org/plugins/decalog/">DecaLog</a>' );
+			if ( class_exists( 'PerfOpsOne\Installer' ) && ! Environment::is_wordpress_multisite() ) {
+				$help .= '<br/><a href="' . esc_url( admin_url( 'admin.php?page=podd-settings&tab=misc&action=install-decalog' ) ) . '" class="poo-button-install"><img style="width:16px;vertical-align:text-bottom;" src="' . \Feather\Icons::get_base64( 'download-cloud', 'none', '#FFFFFF', 3 ) . '" />&nbsp;&nbsp;' . esc_html__('Install It Now', 'device-detector' ) . '</a>';
+			}
 		}
 		add_settings_field(
 			'podd_plugin_options_logger',